[MISB] Factor out common code from the two sensors into abstract base class
