--- sensors/aviation/sensorhub-driver-misb-uas/src/main/java/org/sensorhub/impl/sensor/uas/common/ITimeSynchronizedUasDataProducer.java
@@ -1,13 +1,0 @@
-package org.sensorhub.impl.sensor.uas.common;
-
-import org.sensorhub.api.data.IDataProducer;
-
-/**
- * Common interface implemented by the two sensors in this package. Used in code that needs to work with both.
- */
-public interface ITimeSynchronizedUasDataProducer extends IDataProducer {
-	public void setStreamSyncTime(SyncTime syncTime);
-	public SyncTime getSyncTime();
-	public String getUasFoiUID();
-	public String getImagedFoiUID();
-}
