+++ sensors/aviation/sensorhub-driver-misb-uas/src/main/java/org/sensorhub/impl/sensor/uas/outputs/SensorParams.java
@@ -14,7 +14,8 @@
 package org.sensorhub.impl.sensor.uas.outputs;
 
 import org.sensorhub.api.data.DataEvent;
-import org.sensorhub.impl.sensor.uas.common.ITimeSynchronizedUasDataProducer;
+import org.sensorhub.impl.sensor.uas.UasSensorBase;
+import org.sensorhub.impl.sensor.uas.config.UasConfig;
 import org.sensorhub.impl.sensor.uas.klv.UasDataLinkSet;
 import org.sensorhub.misb.stanag4609.tags.TagSet;
 import org.slf4j.Logger;
@@ -28,7 +29,7 @@ import net.opengis.swe.v20.DataBlock;
  * @author Nick Garay
  * @since Oct. 6, 2020
  */
-public class SensorParams extends UasOutput {
+public class SensorParams<UasConfigType extends UasConfig> extends UasOutput<UasConfigType> {
 
     private static final String SENSOR_OUTPUT_NAME = "sensorParams";
     private static final String SENSOR_OUTPUT_LABEL = "Imaging Sensor Parameters";
@@ -41,7 +42,7 @@ public class SensorParams extends UasOutput {
      *
      * @param parentSensor Sensor driver providing this output
      */
-    public SensorParams(ITimeSynchronizedUasDataProducer parentSensor) {
+    public SensorParams(UasSensorBase<UasConfigType> parentSensor) {
 
         super(SENSOR_OUTPUT_NAME, parentSensor);
